changer

+++ routes/posts.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, render_template, request, flash
+from flask import Blueprint, render_template, request, flash,redirect,url_for
 from flasker.app import db
 from flasker.models import Posts
-from flasker.forms import PostForm
-from flask_login import login_required
+from flasker.forms import PostForm,SearchForm
+from flask_login import login_required,current_user
 
 
 # Create a Blueprint for user routes
@@ -13,11 +13,11 @@ post_bp = Blueprint('post_bp', __name__)
 def add_post():
     form=PostForm()
     if form.validate_on_submit():
-        post=Posts(title=form.title.data,content=form.content.data,author=form.author.data,slug=form.slug.data)
+        poster=current_user.id
+        post=Posts(title=form.title.data,content=form.content.data,poster_id=poster,slug=form.slug.data)
         form.title.data=''
         form.content.data=''
         form.slug.data=''
-        form.author.data=''
         db.session.add(post)
         db.session.commit()
         flash("Poste a été soumis avec succès.")
@@ -38,34 +38,49 @@ def post(id):
 def edit_post(id):
     post=Posts.query.get_or_404(id)
     form=PostForm()
+    id=current_user.id
     if form.validate_on_submit():
         post.title=form.title.data
-        post.author=form.author.data
         post.slug=form.slug.data
         post.content=form.content.data
         db.session.add(post)
         db.session.commit()
         flash('Le poste a été modifié!')
         form.title.data=''
-        form.author.data=''
         form.slug.data=''
         form.content.data=''
-        return redirect(url_for('post_bp.post',id=post.id))
-    form.title.data=post.title
-    form.author.data=post.author
-    form.slug.data=post.slug
-    form.content.data=post.content
-    return render_template('edit_post.html',form=form)
+        return redirect(url_for('post_bp.posts'))
+    if id==post.poster_id:
+        form.title.data=post.title
+        form.slug.data=post.slug
+        form.content.data=post.content
+        return render_template('edit_post.html',form=form)
+    else:
+        flash('Vous ne pouvez modfiier que vos postes')
+        return redirect(url_for('post_bp.posts'))
 
 @post_bp.route('/posts/delete/<int:id>',methods=['GET','POST'])
 @login_required
 def delete_post(id):
     post=Posts.query.get_or_404(id)
-    try:
-        db.session.delete(post)
-        db.session.commit()
-        flash("Le poste a été supprimé avec succès")
-    except:
-        flash("Whoops! Il y a eut un problème.")
+    id=current_user.id
+    if id==post.poster_id:
+        try:
+            db.session.delete(post)
+            db.session.commit()
+            flash("Le poste a été supprimé avec succès")
+        except:
+            flash("Whoops! Il y a eut un problème.")
+    else:
+        flash('Vous ne pouvez supprimer que vos postes')
     posts=Posts.query.order_by(Posts.date_posted)
     return render_template('posts.html',posts=posts)
+
+@post_bp.route('/search',methods=['POST'])
+def search():
+    form=SearchForm()
+    posts=Posts.query
+    if form.validate_on_submit():
+        searched=form.searched.data
+        posts=posts.filter(Posts.content.like('%'+searched+'%')).order_by(Posts.title).all()
+        return render_template("search.html",form=form,searched=searched,posts=posts)
